src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=src/requirements.txt src/requirements/requirements.in
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via requests
dj-database-url==2.1.0
    # via -r src/requirements/requirements.in
django==4.1.11
    # via
    #   -r src/requirements/requirements.in
    #   dj-database-url
    #   djangorestframework
djangorestframework==3.14.0
    # via -r src/requirements/requirements.in
idna==3.4
    # via requests
psycopg2-binary==2.9.7
    # via -r src/requirements/requirements.in
python-decouple==3.8
    # via -r src/requirements/requirements.in
pytz==2023.3.post1
    # via djangorestframework
requests==2.31.0
    # via -r src/requirements/requirements.in
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via dj-database-url
tzdata==2023.3
    # via django
urllib3==2.0.4
    # via requests
